Remove Connection Keep-Alive from request headers
Since version 2.4.0 of requests (https://pypi.python.org/pypi/requests) the
header 'Connection: keep-alive' is sent automatically. According to the API
documentation this should be set to 'Connection: close'. The only way that I
found to set this header was to manually edit the dictionary of headers. This
produces excessive logging when the logging level is set to INFO or lower
because the requests module logs that it is having to re-establish the
connection to the server.

Also, updated the setup.py file so that pip will install the correct version of
requests automatically.